Address review comments
- Merge suggested changes. A separate 'qos' package still needed to be split from 'intdataplane' in order to avoid an import loop.
- Properly attribute and copyright code from the CNI bandwidth plugin
- Add a proper update function for kdd workloads to the felix FV infra
- Adjust test expectations (using 'Consistently()' where appropriate)
- Correctly use '%w' for outputting errors
- Inline calculations used for converting rate and burst in bits {per second,} to TokenBucketState{} values
- Other misc changes
